+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -19,6 +19,7 @@ CHANGELOG
  * Add argument `$buildDir` to `WarmableInterface`
  * Add argument `$filter` to `Profiler::find()` and `FileProfilerStorage::find()`
  * Add `ControllerResolver::allowControllers()` to define which callables are legit controllers when the `_check_controller_is_allowed` request attribute is set
+ * Add `#[MapRequestHeader]` to map header from `Request::$headers`
 
 6.3
 ---